Refactor RequestContext and AuthenticationUtils
