--- app/lib/exceptions.rb
@@ -1,3 +1,0 @@
-module Exceptions
-  class AuthenticationError < StandardError; end
-end
