+++ app/lib/request_context.rb
@@ -1,0 +1,7 @@
+class RequestContext
+  attr_reader :user
+
+  def initialize(args)
+    @user = args[:user]
+  end
+end
