--- app/lib/request_context/context.rb
@@ -1,10 +1,0 @@
-module RequestContext
-  class Context
-
-    attr_reader :user
-
-    def initialize(args)
-      @user = args[:user]
-    end
-  end
-end
